+++ src/index.ts
@@ -3,5 +3,6 @@ export { Signal } from 'signal-polyfill';
 export { action } from './action';
 export { atom, type Atom } from './atom';
 export { computed, type Computed } from './computed';
+export { external } from './external';
 export { get, type Source } from './get';
 export { swap, type Sink } from './swap';